added js

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -3,10 +3,10 @@ package ru.kata.spring.boot_security.demo.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import ru.kata.spring.boot_security.demo.model.Role;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.service.RoleService;
 import ru.kata.spring.boot_security.demo.service.UserService;
@@ -15,11 +15,12 @@ import ru.kata.spring.boot_security.demo.utils.UserIncorrectData;
 
 import javax.validation.Valid;
 import java.security.Principal;
+import java.util.Collection;
 import java.util.List;
 
 
 @RestController
-@RequestMapping("api/")
+@RequestMapping("/api")
 public class AdminController {
     private final UserService userService;
     private final RoleService roleService;
@@ -36,50 +37,39 @@ public class AdminController {
     }
 
     @GetMapping("/admin/user/{id}")
-    public User showAllUsers(@PathVariable ("id") Long id) {
+    public User showUser(@PathVariable ("id") Long id) {
         return userService.getUserById(id);
     }
 
 
-
     @PostMapping("/admin")
-    public String saveNewUser(@ModelAttribute("newUser") @Valid User user, BindingResult bindingResult
-            ,Principal principal, Model model ) {
-        if(bindingResult.hasErrors()) {
-            model.addAttribute("admin", userService.getUserByUsername(principal.getName()));
-            model.addAttribute("allRoles", roleService.getAllRoles());
-            model.addAttribute("allUsers", userService.getAllUsers());
-            model.addAttribute("activeTable", "addUser");
-            return "admin-page";
-        }
+    public User saveNewUser(@RequestBody @Valid User user, BindingResult bindingResult) {
+
         userService.saveUser(user);
-        return "redirect:/admin";
+        return user;
     }
 
-    @PatchMapping("/admin")
-    public String edit(@ModelAttribute("user") @Valid User user, BindingResult bindingResult,
-                                              Model model, Principal principal) {
+    @PutMapping("/admin")
+    public User edit(@RequestBody @Valid User user, BindingResult bindingResult) {
 
-        model.addAttribute("admin", userService.getUserByUsername(principal.getName()));
-        model.addAttribute("allRoles", roleService.getAllRoles());
-        if (bindingResult.hasErrors()) {
-            return "admin-page";
-        }
         userService.saveUser(user);
-        return "redirect:/admin";
+        return user;
     }
 
-    @DeleteMapping("/admin")
-    public String delete(@RequestParam("id") Long id) {
-        userService.deleteUserById(id);
-        return "redirect:/admin";
+    @DeleteMapping("/admin/user/{id}")
+    public String delete(@PathVariable ("id") Long id) {
+
+        if(userService.getUserById(id) != null) {
+            userService.deleteUserById(id);
+
+        }
+        return "User with ID = " + id + " was deleted";
     }
 
-    @ExceptionHandler
-    private ResponseEntity<UserIncorrectData> handleException(
-            NoSuchUserException exception) {
-        UserIncorrectData data = new UserIncorrectData();
-        data.setInfo(exception.getMessage());
-        return new ResponseEntity<>(data, HttpStatus.NOT_FOUND);
+    @GetMapping("/admin/roles")
+    public Collection<Role> getAllRoles(){
+        return roleService.getAllRoles();
     }
+
+
 }
